1.Изменила название класса RestApiPageTest.java на RestApiLinkTest.java (тестируется ссылка в футере jenkins ) 2. Прописала зависимость своих тестов и перенесла тесты из классов SingOut2Test.java и SingIn2Test.java в соответсвующие созданные ранее классы 3. Удалила не нужные классы SingOut2Test.java и SingIn2Test.java
